+++ src/schemas/database/project/expandedProject_v0.ts
@@ -6,13 +6,11 @@ import {
     getStrictSchema,
     getUserWritableSchema,
 } from '../../utility'
-import { BranchTypeDataSchema_v0 } from './branchType_v0'
 import { DeadlineDataSchema_v0 } from './deadline_v0'
 import { DoorTypeDataSchema_v0 } from './doorType_v0'
 import { DuctTypeDataSchema_v0 } from './ductType_v0'
 import { EquipmentDataSchema_v0 } from './equipment_v0'
 import { GraphSchema_v0 } from './graph_v0'
-import { PipeTypeDataSchema_v0 } from './pipeType_v0'
 import { ProjectDataSchema_v0, ProjectPostSchema_v0 } from './project_v0'
 import { RegisterTypeDataSchema_v0 } from './registerType_v0'
 import { RoofTypeDataSchema_v0 } from './roofType_v0'
@@ -28,7 +26,6 @@ import { ZoneDataSchema_v0 } from './zone_v0'
 
 /** Subcollections returned by `expand`. */
 export const ProjectSubcollectionsSchema_v0 = z.object({
-    branchTypes: z.record(z.string(), BranchTypeDataSchema_v0).optional(),
     deadlines: z.record(z.string(), DeadlineDataSchema_v0).optional(),
     doorTypes: z.record(z.string(), DoorTypeDataSchema_v0).optional(),
     ductTypes: z.record(z.string(), DuctTypeDataSchema_v0).optional(),
@@ -37,7 +34,6 @@ export const ProjectSubcollectionsSchema_v0 = z.object({
         z.record(z.string(), APIExportSchema_v0).optional()
     ),
     graph: GraphSchema_v0.optional(),
-    pipeTypes: z.record(z.string(), PipeTypeDataSchema_v0).optional(),
     registerTypes: z.record(z.string(), RegisterTypeDataSchema_v0).optional(),
     roofTypes: z.record(z.string(), RoofTypeDataSchema_v0).optional(),
     sheetFiles: z.record(z.string(), SheetFileDataSchema_v0).optional(),